+++ SampleApp/PayPal-iOS-SDK-Sample-App/ZZFlipsideViewController.m
@@ -13,7 +13,6 @@
 @interface ZZFlipsideViewController ()
 
 @property(nonatomic, strong, readwrite) IBOutlet UISegmentedControl *environmentSegmentedControl;
-@property(nonatomic, strong, readwrite) IBOutlet UISwitch *acceptCreditCardsSwitch;
 @property(nonatomic, strong, readwrite) IBOutlet UITextView *payPalResultTextView;
 @property(nonatomic, strong, readwrite) IBOutlet UILabel *payPalResultLabel;
 @end
@@ -32,7 +31,7 @@
 }
 
 - (void)logEnvironment {
-  NSLog(@"Environment: %@. Accept credit cards? %d", self.delegate.environment, self.delegate.acceptCreditCards);
+  NSLog(@"Environment: %@.", self.delegate.environment);
 }
 
 - (void)viewWillAppear:(BOOL)animated {
@@ -46,11 +45,6 @@
       break;
     }
   }
-  self.acceptCreditCardsSwitch.on = self.delegate.acceptCreditCards;
-
-#if !HAS_CARDIO
-  self.acceptCreditCardsSwitch.enabled = NO;
-#endif
   
   if ([self.delegate resultText]) {
     NSLog(@"%@", [self.delegate resultText]);
@@ -73,11 +67,6 @@
   if (self.environmentSegmentedControl == sender) {
     [self.delegate setPayPalEnvironment:[[self.environmentSegmentedControl titleForSegmentAtIndex:self.environmentSegmentedControl.selectedSegmentIndex] lowercaseString]];
   }
-  [self logEnvironment];
-}
-
-- (IBAction)processCreditCardsChanged:(id)sender {
-  [self.delegate setAcceptCreditCards:self.acceptCreditCardsSwitch.on];
   [self logEnvironment];
 }